move logic to controller
