+++ src/config/config.js
@@ -3,5 +3,6 @@ var config = {};
 config.adminAPI = "http://localhost:8080/admin/";
 config.apiMetaDataAPI = "http://localhost:9090/apiMetadata/"
 config.port = 3000
+config.mode = 'single'
 
 module.exports = config;
